+++ BepisDynamicVariablePowerTools/Patches/RenameDynVarSources.cs
@@ -4,7 +4,6 @@ using FrooxEngine.FrooxEngine.ProtoFlux.CoreNodes;
 using FrooxEngine.ProtoFlux;
 using FrooxEngine.UIX;
 using HarmonyLib;
-using System.Reflection;
 
 namespace BepisDynamicVariablePowerTools.Patches;
 
@@ -21,57 +20,25 @@ public static class RenameDynvarSources
 
         Type nodeType = node.GetType();
         Type baseType = nodeType.IsGenericType ? nodeType.GetGenericTypeDefinition() : nodeType;
-        Type innerType = nodeType.IsGenericType ? nodeType.GenericTypeArguments[0] : nodeType;
 
-        if (baseType == typeof(ValueSource<>))
+        if (baseType == typeof(ValueSource<>) || baseType == typeof(ObjectValueSource<>) || baseType == typeof(ReferenceSource<>))
         {
-            typeof(RenameDynvarSources).GetGenericMethod("HandleValueSource", BindingFlags.Static | BindingFlags.Public, innerType).Invoke(null, [__instance, node]);
-        }
-        if (baseType == typeof(ReferenceSource<>))
-        {
-            typeof(RenameDynvarSources).GetGenericMethod("HandleReferenceSource", BindingFlags.Static | BindingFlags.Public, innerType).Invoke(null, [__instance, node]);
-        }
-    }
+            IValue sourceTarget = (IValue)Traverse.Create(node).Property<ISyncRef>("RootSourceReference").Value.Target;
 
-    public static void HandleValueSource<T>(ProtoFluxNodeVisual visual, ValueSource<T> source) where T : unmanaged
-    {
-        var refValue = (IValue<T>)source.RootSourceReference.Target;
-        var parent = refValue.Parent;
-        if (parent is Component c)
-        {
-            var compType = c.GetType();
-            if (compType.IsGenericType)
+            var fieldName = sourceTarget.Name;
+            var parent = sourceTarget.Parent;
+            var field = Traverse.Create(parent).Field("VariableName");
+            if (field.FieldExists())
             {
-                if (compType == typeof(DynamicValueVariable<T>))
-                {
-                    string variableName = Traverse.Create(parent).Field<Sync<string>>("VariableName").Value;
-                    ReplaceNodeText(visual, c, variableName);
-                }
+                string variableName = field.GetValue<Sync<string>>().Value;
+                ReplaceNodeText(__instance, (Component)parent, variableName, fieldName);
             }
         }
     }
 
-    public static void HandleReferenceSource<T>(ProtoFluxNodeVisual visual, ReferenceSource<T> source) where T : class, IWorldElement
-    {
-        var refValue = (SyncRef<T>)source.RootSourceReference.Target;
-        var parent = refValue.Parent;
-        if (parent is Component c)
-        {
-            var compType = c.GetType();
-            if (compType.IsGenericType)
-            {
-                if (compType == typeof(DynamicReferenceVariable<T>))
-                {
-                    string variableName = Traverse.Create(parent).Field<Sync<string>>("VariableName").Value;
-                    ReplaceNodeText(visual, c, variableName);
-                }
-            }
-        }
-    }
-
-    public static void ReplaceNodeText(ProtoFluxNodeVisual visual, Component component, string VariableName)
+    public static void ReplaceNodeText(ProtoFluxNodeVisual visual, Component component, string VariableName, string FieldName)
     {
         Slot holder = visual.Slot.FindChildInHierarchy("Overlapping Layout").FindChild("Horizontal Layout").Children.First();
-        holder.GetComponent<Text>().Content.Value = $"<size=60%>Dynamic Variable:</size>\n<i>{VariableName}</i><size=75%>\n(On: {component.Slot.Name})</size>";
+        holder.GetComponent<Text>().Content.Value = $"<size=60%>Dynamic Variable: ({FieldName})</size>\n<i>{VariableName}</i><size=75%>\n(On: {component.Slot.Name})</size>";
     }
 }